+++ gallery/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2
+- Minor improvements in the dialog functions.
+- Removed the "Remember" selection; FastJS will now remember your selection(s) by default.
+- Added a playlist feature to manage multiple scripts more efficiently.
+- Updated the configuration view to display only the filename (e.g., `script.js`) instead of the full path for each selected script.
+
 ## 1.1
 - Updated the File Browser to open in the Apps/Scripts folder